Import amount of overdraft into agios

+++ src/main/java/org/example/agios/Agios.java
@@ -1,12 +1,14 @@
 package org.example.agios;
 
+import static org.example.withdraw.Withdraw.overdraftAmount;
+
 public class Agios {
 
     public static void DisplayAgios(){
         System.out.println("Agios");
     }
 
-    public static double CalculAgios(int overdraft, int dayOfOverdraft){
+    public static double CalculAgios(double overdraft, int dayOfOverdraft){
         double DOverdraft = overdraft;
         System.out.println("Overcraft : "+DOverdraft+" and Number of day of overdraft : "+dayOfOverdraft+".");
         return DOverdraft;
